riscv: enable SBI RFENCE call support

+++ hypervisor/include/arch/riscv/asm/smp.h
@@ -11,7 +11,7 @@
 #include <asm/current.h>
 
 struct swi_vector {
-	uint32_t type;
+	uint64_t type;
 	uint64_t param;
 };
 #endif /* !__ASSEMBLY__ */
@@ -23,7 +23,7 @@ extern void stop_cpu(void);
 extern void smp_init_cpus(void);
 extern void smp_clear_cpu_maps (void);
 extern void start_pcpus(void);
-extern void send_single_swi(uint16_t pcpu_id, uint32_t vector);
+extern void send_single_swi(uint16_t pcpu_id, uint64_t vector);
 extern int kick_pcpu(int cpu);
 extern int do_swi(int cpu);
